Title: 비교에 따른 연산, Time: 115ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240122/비교에 따른 연산/operation-based-on-comparison.py
@@ -5,4 +5,4 @@ b=int(x[1])
 if a>b:
     print(a*b)
 else:
-    print(b/a)
+    print(int(b/a))
